% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contr.prev.R
\name{contr.prev}
\alias{contr.prev}
\title{Additional Contrasts Matrices}
\usage{
contr.prev(n, contrasts = TRUE)
}
\arguments{
\item{n}{a vector of levels for a factor, or the number of levels}

\item{contrasts}{a logical indicating whether contrasts should be computed
\code{\link[Matrix]{dgCMatrix-class}}), using package \pkg{Matrix}.}
}
\value{
A matrix with \code{n} rows and \code{k} columns, with \code{k=n-1}
if contrasts is \code{TRUE} and \code{k=n} if contrasts is \code{FALSE}.
}
\description{
Return a matrix of contrasts.
}
\details{
\code{contr.prev} returns contrasts which contrast each level
starting from the second with the previous level. Obviously this makes sense
rather for an ordered factor. Be aware that these contrasts are \bold{not}
orthogonal (and consequently are unsuitable to be used with \code{\link[stats]{aov}}).
}
\examples{
contr.prev(4)
}
\seealso{
\code{\link[stats]{contrast}}
}
